requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
httpx>=0.27.0

# STT
faster-whisper>=1.0.0

# GLaDOS TTS (ONNX inference + IPA phonemisation)
onnxruntime>=1.17.0
phonemizer>=3.2.1
numpy>=1.24.0
soundfile>=0.12.1

# ATLAS TTS (Piper)
piper-tts>=1.2.0

# Model download helper
huggingface-hub>=0.24.0
